Split queue and router
