+++ app/src/main/res/layout/queue.xml
@@ -7,22 +7,6 @@
     android:layout_height="match_parent"
     android:background="@color/background">
 
-    <ProgressBar
-        android:id="@+id/QueueProgressBar"
-        style="?android:attr/progressBarStyle"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="8dp"
-        android:layout_marginEnd="8dp"
-        android:layout_marginLeft="8dp"
-        android:layout_marginRight="8dp"
-        android:layout_marginStart="8dp"
-        android:layout_marginTop="8dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
     <TextView
         android:id="@+id/QueueServerText"
         android:layout_width="100dp"
@@ -130,7 +114,7 @@
         app:layout_constraintHorizontal_bias="0.0"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/QueueViewSearch" />
+        app:layout_constraintTop_toBottomOf="@+id/QueueViewSearchButton" />
 
 
 </android.support.constraint.ConstraintLayout>
