+++ app/src/main/res/layout/search.xml
@@ -13,21 +13,20 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginBottom="8dp"
-        android:layout_marginEnd="8dp"
         android:layout_marginLeft="8dp"
         android:layout_marginRight="8dp"
-        android:layout_marginStart="8dp"
         android:layout_marginTop="8dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+        app:layout_constraintBottom_toBottomOf="@+id/SearchView"
+        app:layout_constraintLeft_toLeftOf="@+id/SearchView"
+        app:layout_constraintRight_toRightOf="@+id/SearchView"
+        app:layout_constraintTop_toTopOf="@+id/SearchView"
+        app:layout_constraintVertical_bias="0.14999998" />
 
     <TextView
         android:id="@+id/SearchServerText"
         android:layout_width="100dp"
         android:layout_height="28dp"
-        android:layout_marginLeft="7dp"
+        android:layout_marginLeft="8dp"
         android:layout_marginTop="8dp"
         android:text="Server Key: "
         android:textColor="@color/white"
