+++ app/src/main/res/values/styles.xml
@@ -41,7 +41,4 @@
         <item name="android:textColor">#6de873</item>
     </style>
 
-
-
-
 </resources>
